Move couch node system service to libs
