+++ src/services/monitor.ts
@@ -2,7 +2,7 @@ import * as Effect from 'effect/Effect';
 import * as Context from 'effect/Context';
 import { CouchDbInfo, getDbsInfoByName } from '../libs/couch/dbs-info';
 import { CouchDesignInfo, getDesignInfo } from '../libs/couch/design-info';
-import { CouchNodeSystem, getCouchNodeSystem } from './couch/node-system';
+import { CouchNodeSystem, getCouchNodeSystem } from '../libs/couch/node-system';
 import { Array, Clock, Number, Option, pipe } from 'effect';
 import { LocalDiskUsageService } from './local-disk-usage';
 import { ResponseError } from '@effect/platform/HttpClientError';
