+++ test/services/monitor.spec.ts
@@ -2,8 +2,8 @@ import { afterEach, describe, it } from 'mocha';
 import { Effect, Either, Layer, Option, TestClock } from 'effect';
 import { expect } from 'chai';
 import { MonitorService } from '../../src/services/monitor';
-import * as CouchNodeSystemLib from '../../src/services/couch/node-system';
-import { CouchNodeSystem } from '../../src/services/couch/node-system';
+import * as CouchNodeSystemLib from '../../src/libs/couch/node-system';
+import { CouchNodeSystem } from '../../src/libs/couch/node-system';
 import sinon, { SinonStub } from 'sinon';
 import * as CouchDbsInfo from '../../src/libs/couch/dbs-info';
 import { CouchDbInfo } from '../../src/libs/couch/dbs-info';
